OMP thread time & variable num elements argument
quicksortOMP.c is now run thru a command line argument, which is now included in the qs-job slurm script. It also uses omp_get_wtime() to time the sorting instead of using CLI time command to time the entire program to more accurately time the sorting itself. The output slurm file is included.